Adding auto shutdown functionality (#36)
As the title suggests, adding auto shutdown when no one is on the server for 15 minutes. As a side effect, there was also some cleanup of code to shutdown the server so that now all shutdown attempts will also upload the world and close the stack. Testing was done locally with one player, however testing on AWS with multiple people would probably be useful.