Adding tests and testing CI (#46)
Adding tests which can be run using the npm test command. I ended up picking Ava and Sinon for the testing framework because it seemed new and if the project continues as we have been it will be mature by the time we are ready for release. (ha) Also added to the pull request workflow set up in the previous pull request to have the tests run on all PRs.