Adding flavor config file support (#47)
In order to support the fact that modded servers need to sync additional files, the ability to include an autostack-config.json file as part of the flavor files has been added. At the moment, this file just supplies the syncFolders and syncFiles that will be saved to s3, but as discussed could potentially included additional configuration information in the future. This config files is currently optional and if not give for a flavor will default to the regular Minecraft configuration.