Moving Node package to be entire repo (#56)
Essentially just move packages.json and related config files up a directory in the repo. As per #51, this will allow us to lint and prettify all Javascript files we have. For the moment, the client portal code is excluded from linting as the linter needs a different config file to handle browser Javascript. While adjusting the configuration files, I was able to consolidate some of them as they were no longer needed as well.